requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
python-multipart
notion-client